Remove bootstrap scaffold row/span classes.
Apparently I don't know enough about that stuff to get it working
together with the animations.  Spent a couple of hours trying to track
down an unwanted jump of the task view after the fade in.  Tracked it
down the the bootstrap classes.  Suspect it has something to do with
margins and/or padding, but...