Keep forms from occupying the same space (even when invisible).
The controls on the visible form did not work correctly when
there was a non-visible form beneath it.  I could give focus to a text
box with the mouse, I had to tab to.  And the buttons didn't work,
etc.  So I just changed it so that the forms don't moved into the
shared space until we are switching screens.